+++ test/index.spec.js
@@ -100,7 +100,7 @@ test('bissle >> route have no defined ID', t => {
 test('bissle >> exposed query schema', t => {
   const { server } = setup();
 
-  t.deepEqual(_.keys(server.plugins.bissle.schema).sort(), ['page', 'per_page']);
-  t.equal(server.plugins.bissle.schema.page.isJoi, true);
+  t.deepEqual(_.keys(server.plugins.bissle.scheme).sort(), ['page', 'per_page']);
+  t.equal(server.plugins.bissle.scheme.page.isJoi, true);
   t.end();
 });