+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # master
 
 ### unreleased
+- **fix:** flaws
 
 ### [v0.1.2] - 2016-06-25
 - **fix:** fix flaws