+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # master
 
 ### unreleased
+
 - **fix:** flaws
 - **fix:** errors line breaks
 - **fix:** flaws
 - **fix:** change max-len eslint rule
 - **fix:** wrong tests
 - **fix:** missing left query parameters in HAL navigation links
+- **fix:** flaws
 
 ### [v0.1.2] - 2016-06-25
 - **fix:** fix flaws